+++ CHANGELOG.md
@@ -49,14 +49,52 @@ available on the `main` branch.
 
 ### Added
    
- - Added `grav_undef()` to undo gravitational bending of the observed light to obtain nominal astrometric positions
+ - New `novas_timespec` structure for the self-contained definition of precise astronomical time (`timescale.c`). You 
+   can set the time via `novas_set_time()` or `novas_set_split_time()` to a JD date in the timescale of choice (UTC, 
+   UT1, GPS, TAI, or TT), or to a UNIX time with `novas_set_unix_time()`. Once set, you can obtain an expression of 
+   that time in any timescale of choice via `novas_get_time()`, `novas_get_split_time()` or `novas_get_unix_time()`.
+   And, you can create a new time specification by incrementing an existing one, using `novas_increment_time()`, or 
+   measure time differences via `novas_diff_time()`. 
+   
+ - New observing-frame based approach for calculations (`frames.c`). A `novas_frame` object uniquely defines both the 
+   place and time of observation, with a set of pre-calculated transformations and constants. Once the frame is 
+   defined it can be used very efficiently to calculate positions for multiple celestial objects with minimal 
+   additional computational cost. The frames API is also more elegant and simpler than the low-level NOVAS C approach 
+   for performing the same kind of calculations. And, frames are inherently thread-safe since post-creation their 
+   internal state is never modified during the calculations. The following new functions were added: 
+   `novas_make_frame()`, `novas_change_observer()`, `novas_geom_posvel()`, `novas_sky_pos()`, `novas_app_to_hor()`, 
+   `novas_app_to_geom()`, `novas_hor_to_app()`, `novas_make_transform()`, `novas_invert_transform()`, 
+   `novas_transform_vector()`, and `novas_transform_sky_pos()`.
+   
+ - Added `grav_undef()` to undo gravitational bending of the observed light to obtain geometric positions from
+   observed ones.
  
- - Added new observer locations `NOVAS_SOLAR_SYSTEM_OBSERVER` (for spacecraft orbiting the Sun), and 
-   `NOVAS_OBSERVER_AT_BARYCENTER` for an imaginary observe at the Solar System Baricenter.
+ - Added new observer locations `NOVAS_AIRBORNE_OBSERVER` for an observer moving relative to the surface of Earth e.g.
+   in an aircraft or balloon based telescope platform, and `NOVAS_SOLAR_SYSTEM_OBSERVER` for spacecraft orbiting the 
+   Sun.
    
  - Added coordinate reference systems `NOVAS_MOD` (Mean of Date) which includes precession by not nutation and
    `NOVAS_J2000` for the J2000 dynamical reference system.
 
+ - New observer locations `NOVAS_AIRBORNE_OBSERVER` and `NOVAS_SOLAR_SYSTEM_OBSERVER`, and corresponding
+   `make_airborne_observer()` and `make_solar_system_observer()` functions. Airborne observers have an earth fixed
+   momentary location, defined by longitude, latitude, and altitude, the same was as for a stationary observer on
+   Earth, but are moving relative to the surface, such as in an aircraft or balloon observatory. Solar-system
+   observers are similar to observers in Earth-orbit but their momentary position and velocity is defined relative
+   to the Solar System Barycenter, instead of the geocenter.
+
+ - Added humidity field to `on_surface` structure, e.g. for refraction calculations at radio wavelengths. The
+   `make_on_surface()` function will set humidity to 0.0, but the user can set the field appropriately afterwards.
+
+ - New set of built-in refraction models to use with the frame-based `novas_app_to_hor()` function. The models
+   `novas_standard_refraction()` and `novas_optical_refraction()` implement the same refraction model as `refract()` 
+   in NOVAS C 3.1, with `NOVAS_STANDARD_ATMOSPHERE` and `NOVAS_WEATHER_AT_LOCATION` respectively, including the 
+   reversed direction provided by `refract_astro()`. The user may supply their own refraction models to
+   `novas_app_to_hor()` also, and may make used of the generic reversal function `novas_inv_refract` to calculate
+   refraction in the reverse direction (observer vs astrometric elevations) as needed.
+
+ - Added radio refraction model `novas_radio_refraction()` based on the formulae by Berman &amp; Rockwell 1976.
+ 
  - `make help` to provide a brief list and explanation of the available build targets. (Thanks to `@teuben` for 
    suggesting this.)
 
@@ -69,6 +107,8 @@ available on the `main` branch.
  - Improved precision of some calculations, like `era()`, `fund_args()`, and `planet_lon()` by being more careful
    about the order in which terms are accumulated and combined, resulting in a small improvement on the few uas 
    (micro-arcsecond) level.
+   
+ - The `ra` or `dec` arguments passed to `vector2radec()` may now be NULL if not required.
 
  - The default make target is now `distro`. It's similar to the deprecated `api` target from before except that it 
    skips building `static` libraries.